feat(virtq): cleanup send + sync bounds
